Add data dictionary
Added data dictionary which can be used to factorise the CSV files when they're loaded in for plotting. Usage will depend on the plotting code, I'll amend appropriately shortly.

Structure of the dictionary:
Named list with name = stratification variable (e.g. sex) and values = a vector of strings that correspond to the factor levels in the right order