relabel current summary stats in the data for clarity
Need to support splitting out the latest observation for each participant from the rest of the data. This is currently achieved by duplicating the record and setting a dummy date in the future. This commit aims to improve that by setting date for these duplicate records to NA and marking them in a new variable. This is more consistent and avoids potential double counting or unintended consequences from using max(date).